data/ingestion_pipeline/requirements.txt
Flask==2.3.2
requests==2.31.0
google-cloud-storage==2.14.0
